+++ packages/twenty-server/src/modules/workflow/workflow-runner/workflow-runner.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 
-import { WorkflowRunService } from 'src/modules/workflow/workflow-runner/services/workflow-run.service';
-import { WorkflowRunJob } from 'src/modules/workflow/workflow-runner/jobs/workflow-run.job';
+import { WorkflowRunnerService } from 'src/modules/workflow/workflow-runner/workflow-runner.service';
+import { WorkflowRunnerJob } from 'src/modules/workflow/workflow-runner/workflow-runner.job';
 import { WorkflowCommonModule } from 'src/modules/workflow/common/workflow-common.module';
 import { ServerlessFunctionModule } from 'src/engine/metadata-modules/serverless-function/serverless-function.module';
 
 @Module({
   imports: [WorkflowCommonModule, ServerlessFunctionModule],
-  providers: [WorkflowRunService, WorkflowRunJob],
+  providers: [WorkflowRunnerService, WorkflowRunnerJob],
 })
 export class WorkflowRunnerModule {}
